feat:创建课程

+++ app/src/main/java/com/yuebing/aicoursesystemandroid/model/User.java
@@ -7,7 +7,7 @@ import lombok.Data;
 public class User {
     private Integer id;
 
-    private Integer userid;
+    private Long userid;
 
     private String username;
